--- generator/wallet/xfsc_wallet.py
@@ -1,8 +1,0 @@
-from generator.wallet.wallet import WalletConnector
-
-
-class XFSCWallet(WalletConnector):
-    """
-    Abstraction XFSC wallet, called Organization Credential Manager.
-    See https://projects.eclipse.org/projects/technology.xfsc
-    """
