+++ requirements.txt
@@ -4,4 +4,4 @@ PyYAML==6.*
 click==8.1.7
 pyshacl==0.25.0
 linkml==1.6.7
-
+rdflib==7.0.0
