--- sd/example.jsonld
@@ -1,14 +1,0 @@
-{
-    "@context": { "@vocab": "http://schema.org/" },
-    "@id": "http://example.org/ns#Bob",
-    "@type": "Person",
-    "givenName": "Robert",
-    "familyName": "Junior",
-    "birthDate": "1961-07-07",
-    "deathDate": "1968-09-10",
-    "address": {
-        "@id": "http://example.org/ns#BobsAddress",
-        "streetAddress": "1600 Amphitheatre Pkway",
-        "postalCode": 10404
-    }
-}
