--- sd/gx_service_offering_example.jsonld
@@ -1,65 +1,0 @@
-{
-  "@context": {
-    "xsd": "http://www.w3.org/2001/XMLSchema#",
-    "gx": "https://https://registry.lab.gaia-x.eu//v1$/gx#",
-    "ex": "http://example.org/"
-  },
-  "@id": "ex:Compute",
-  "@type": "gx:ServiceOffering",
-  "gx:providedBy": {
-    "@id": "did:example:123",
-    "@type": "gx:LegalParticipant",
-    "gx:legalAddress": {
-      "@type": "gx:legalAddress",
-      "gx:countrySubdivisionCode": {
-        "@value": "DE-BE",
-        "@type": "xsd:string"
-      }
-    },
-    "gx:headquarterAddress": {
-      "@type": "gx:legalAddress",
-      "gx:countrySubdivisionCode": {
-        "@value": "DE-BE",
-        "@type": "xsd:string"
-      }
-    },
-    "gx:legalRegistrationNumber": {
-      "@type": "gx:legalRegistrationNumber",
-      "gx:leiCode": {
-        "@value": "123",
-        "@type": "xsd:string"
-      },
-      "gx:vatID": {
-        "@value": "123",
-        "@type": "xsd:string"
-      },
-      "gx:EORI": {
-        "@value": "123",
-        "@type": "xsd:string"
-      },
-      "gx:EUID": {
-        "@value": "123",
-        "@type": "xsd:string"
-      },
-      "gx:taxID": {
-        "@value": "123",
-        "@type": "xsd:string"
-      }
-    }
-  },
-  "gx:policy": "IaaS",
-  "gx:termsAndConditions": {
-    "@type": "gx:SOTermsAndConditions",
-    "gx:URL": {
-      "@value": "https://www.acme.com/compute/tac",
-      "@type": "xsd:string"
-     },
-    "gx:hash": "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
-  },
-  "gx:dataAccountExport": {
-    "@type": "gx:dataAccountExport",
-    "gx:requestType": "API",
-    "gx:accessType": "digital",
-    "gx:formatType": "application/json"
-  }
-}
